+++ cogs/core.py
@@ -1,41 +1,11 @@
 from discord.ext import commands
 from discord.utils import get
 import discord, traceback, json, datetime, inspect
+from ext.imports_share import chk_sudo
 BOTSETFILE = "ext/bot_settings.json"
 SETFILE = "data/settings.json"
 settings = json.load(open(SETFILE, 'r'))
 stringTable = json.load(open('ext/wrds.json', 'r'))
-
-
-def is_sudoers(member: discord.Member):
-    """\
-    Type: function
-    Checks if the provided member has admin roles (has moderating priviledges)
-    This function fetches the Admin roles list from the settings `dict()`
-    ---
-    return: bool
-    """
-    if member.guild.owner == member:
-        return True
-    for role in member.roles:
-        try:
-            if role.name in settings[f'g{member.guild.id}']['sudoers']:
-                return True
-        except KeyError:
-            settings[f'g{member.guild.id}'] = {"sudoers": [], "prefix": ["/"]}
-            with open(SETFILE, 'w') as outfile:
-                json.dump(settings, outfile)
-    return False
-
-
-def chk_sudo():
-    """\
-    Type: decorator
-    The command will only be able to be executed by the author if the author is owner or have permissions
-    """
-    async def predicate(ctx):
-        return is_sudoers(ctx.author)
-    return commands.check(predicate)
 
 
 class Core(commands.Cog):
@@ -161,6 +131,8 @@ class Core(commands.Cog):
         e = discord.Embed(title='Command list', description='wd: <GLOBAL>')
         count = 1
         for cmd in all_cmds:
+            if cmd.hidden:
+                continue
             e.add_field(name=cmd.name, value=cmd.help or "<no help>")
             count += 1
         await ctx.send(embed=e)